refactoring current structure
