+++ assets/css/index.css
@@ -1,6 +1,5 @@
 @import "tailwindcss";
 
-
 @theme {
     --animate-fade-in: fade-in 150ms ease-in-out;
     --animate-fade-out: fade-out 150ms ease-in-out;
@@ -29,10 +28,6 @@
 }
 
 :root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-
   color-scheme: light dark;
   color: rgba(255, 255, 255, 0.87);
   background-color: #242424;
@@ -73,11 +68,6 @@ h1 {
   a:hover {
     color: #747bff;
   }
-
-  /* NOTE: ORIGINAL (MIT PETZI ABER NOCHMAL ABKLÄREN, OB STYLES SO PASSEN - LAUT VS CODE HAB ICH (LUKAS) DIE STYLES EINGEFÜGT KANN MICH ABER NICHT DRAN ERINNERN DAS JE GETAN ZU HABEN */
-  /* button {
-    background-color: #f9f9f9;
-  } */
 
   /* NOTE: VORÜBERGEHENED */
   button {
