+++ src/App.jsx
@@ -10,14 +10,14 @@ import {Car} from "./components/Car";
 import {Policeman} from "./components/Policeman";
 import {Streetbay} from "./components/Streetbay";
 
-import './App.css'
+import './../assets/css/App.css'
 import {generateUUID, randInt} from "three/src/math/MathUtils.js";
 import {generateCarAndDriverProfile} from "./utils/generateCarAndDriverProfile.js";
 
 import {Startmenu} from "./components/Startmenu";
 import {CarControlTextbox} from "./components/textboxes/CarControlTextbox";
 import {CarAndDriverProfileTextbox} from "./components/textboxes/CarAndDriverProfileTextbox";
-import {DocumentManager} from "./components/DocumentManager";
+import {DocumentManager} from "./components/manager/DocumentManager";
 
 
 function App() {
@@ -34,11 +34,11 @@ function App() {
     // ################### REFERENCES ###################
     const carRefs = useRef({});
 
-    // function executes when car is selected (onSelect)
     function handleSelectedCar (carObject) {
         setSelectedCar(carObject);
     }
 
+    // Todo: auslagern in Car.jsx
     // spawns new car
     useEffect(() => {
         let respawnTime = randInt(2000, 5000);
@@ -56,7 +56,7 @@ function App() {
         return () => clearInterval(intervalId);
     }, [addCar]);
 
-
+    // Todo: auslagern in Car.jsx
     // effect for making car (not) selectable
     useEffect(() => {
         if (!selectedCar) return;
@@ -76,7 +76,7 @@ function App() {
     return (
         <div className={`h-full ${hoveringCar ? 'cursor-pointer' : ''}` }>
             <Canvas camera={{position: [7, 14, -16], fov: 70}}>
-                {/* GAME COMPONENTS */}
+                {/* UTIL COMPONENTS */}
                 <axesHelper/>
                 <OrbitControls/>
 
@@ -91,7 +91,6 @@ function App() {
 
                 {cars.map((car) => {
                     // if no refference on a car id in the "cars" store (store.js) exists, create a new reference to that id
-                    // NOTE: Stimmt dieser Kommentar?
                     if (!carRefs.current[car.id]) {
                         carRefs.current[car.id] = createRef();
                     }
@@ -119,9 +118,11 @@ function App() {
                         onClose={() => setSelectedCar(null)}
                     />
 
+                    // Todo: -3 ist die z-Position vom Policeman, müsste ausgelagert werden in eine Config Datei
                     {/* NOTE: -3 ist die Z-Koordinate des Autos, wenn es hält (siehe 'CatmullRomCurve3' in Car.jsx) */}
                     {stoppedCar && stoppedCar.position.z === -3 && (
                         <>
+                            // Todo: Brauchen wir diese Box in Zukunft? Soll sich der Spieler die Infos merken?
                             <CarAndDriverProfileTextbox
                                 selectedCar={selectedCar}
                                 stoppedCar={cars.find(car => car.stopped)}
@@ -130,7 +131,6 @@ function App() {
                             {(stoppedCar.handedOutDriversLicense || stoppedCar.handedOutVehicleDocuments) && (
                                 <DocumentManager selectedCar={selectedCar} />
                             )}
-
                         </>
                     )}
                 </>
