+++ src/utils/streetbayEntries/streetbayEntry.js
